Revert moving Pydantic model definitions into models.py file
Having Pydantic model definitions in their own file and importing them causes issues with Sphinx multiversion documentation builds. For example, qiskit-on-iqm has iqm-client as a dependency. When building old versions (e.g. 1.0) of documentation of qiskit-on-iqm, Sphinx uses matching version of qiskit-on-iqm, but also the latest version iqm-client, so qiskit-on-iqm fails to find the model definitions for import
